requirements.txt
streamlit
torch>=2.6.0
transformers
Pillow
requests
pyarrow
